\name{AUCopt-package}
\alias{AUCopt-package}
\alias{AUCopt}
\docType{package}
\title{
\packageTitle{AUCopt}
}
\description{
\packageDescription{AUCopt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AUCopt}
\packageIndices{AUCopt}


}
\author{
\packageAuthor{AUCopt}

Maintainer: \packageMaintainer{AUCopt}
}
\references{
RcppArmadillo
}

\keyword{AUC, optimization}
