\name{AUCopt-package}
\alias{AUCopt-package}
\alias{AUCopt}
\docType{package}
\title{
\packageTitle{AUCopt}
}
\description{
\packageDescription{AUCopt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AUCopt}
\packageIndices{AUCopt}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{AUCopt}

Maintainer: \packageMaintainer{AUCopt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
